requirements.txt
pyyaml
numpy
scipy
torch==2.6.0
torchvision
pillow
rouge==1.0.1
transformers>=4.49.0
qwen-vl-utils
tiktoken
openai
anthropic
google-generativeai
datasets
tenacity
transformers_stream_generator
matplotlib
einops
accelerate
huggingface-hub